requirements.txt
pydantic>=2.0.0
typer[all]>=0.9.0
rich>=13.0.0
python-frontmatter>=1.0.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
pytest>=7.0.0
openai>=1.0.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.20.0